+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# v0.1.7
+# v0.1.7 (2014-06-02)
 
 - Update the `toThrowWith` matcher to accept a pattern to match the message
+- Update the `toThrowWith` matcher to accept the expected class of a thrown exception
+- Add a `toBeAnInstanceOf(type)` matcher
 - Add `toBeTrue` and `toBeFalse` matchers
+
+The `toBeAnInstanceOf` matcher and the `anInstanceOf` parameter have been added because Dart2JS does not fully implement mirrors (in particular, it does not support `isSubtypeOf`, which is used by `toBeA` and `toThrowWith(type:)`), the `toBeAnInstanceOf` matcher and the `anInstanceOf` parameter have been added. See API docs for more information.
 
 # v0.1.6 (2014-05-28)